+++ lib/code-caller/code-caller-docker.js
@@ -13,10 +13,10 @@ const execa = require('execa');
 const bindMount = require('@prairielearn/bind-mount');
 const { instrumented } = require('@prairielearn/opentelemetry');
 
-const config = require('./config');
-const logger = require('./logger');
+const config = require('../config');
+const logger = require('../logger');
 const { FunctionMissingError } = require('./code-caller-shared');
-const dockerUtil = require('./dockerUtil');
+const dockerUtil = require('../dockerUtil');
 const { AsyncResource } = require('async_hooks');
 
 /** @typedef {typeof CREATED | typeof WAITING | typeof IN_CALL | typeof EXITING | typeof EXITED} CallerState */
@@ -499,7 +499,7 @@ class DockerCaller {
     this.debug('enter _callIsFinished()');
     if (!this._checkState([IN_CALL])) return;
     this._clearTimeout();
-    /** @type {import('../executor').Results} */
+    /** @type {import('../../executor').Results} */
     let data = null;
     let err = null;
     try {
