+++ lib/code-caller/code-caller-python.js
@@ -8,7 +8,7 @@ const { AsyncResource } = require('async_hooks');
 const util = require('util');
 
 const { FunctionMissingError } = require('./code-caller-shared');
-const logger = require('./logger');
+const logger = require('../logger');
 
 const CREATED = Symbol('CREATED');
 const WAITING = Symbol('WAITING');
