+++ lib/code-caller/index.js
@@ -6,11 +6,12 @@ const { v4: uuidv4 } = require('uuid');
 const debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
 const util = require('util');
 
-const logger = require('./logger');
-const config = require('./config');
-const load = require('./load');
-const { DockerCaller } = require('./code-caller-docker');
+const logger = require('../logger');
+const config = require('../config');
+const load = require('../load');
+const { DockerCaller, init: initCodeCallerDocker } = require('./code-caller-docker');
 const { PythonCaller } = require('./code-caller-python');
+const { FunctionMissingError } = require('./code-caller-shared');
 
 /**
  * This module maintains a pool of PythonCaller workers, which are used any
@@ -28,12 +29,16 @@ const { PythonCaller } = require('./code-caller-python');
 let pool = null;
 
 module.exports = {
-  init() {
+  async init() {
     debug('init()');
 
     const { workersExecutionMode } = config;
     if (workersExecutionMode !== 'container' && workersExecutionMode !== 'native') {
       throw new Error(`unknown config.workersExecutionMode: ${workersExecutionMode}`);
+    }
+
+    if (config.workersExecutionMode === 'container') {
+      await initCodeCallerDocker();
     }
 
     const numWorkers = config.workersCount ?? Math.ceil(config.workersPerCpu * os.cpus().length);
@@ -156,4 +161,6 @@ module.exports = {
       return fnResult;
     }
   },
+
+  FunctionMissingError,
 };
