+++ tests/helperServer.js
@@ -18,7 +18,7 @@ const syncFromDisk = require('../sync/syncFromDisk');
 const freeformServer = require('../question-servers/freeform');
 const cache = require('../lib/cache');
 const localCache = require('../lib/local-cache');
-const codeCallers = require('../lib/code-callers');
+const codeCaller = require('../lib/code-caller');
 const externalGrader = require('../lib/externalGrader');
 const externalGradingSocket = require('../lib/externalGradingSocket');
 
@@ -100,10 +100,9 @@ module.exports = {
             load.initEstimator('python', 1);
             callback(null);
           },
-          function (callback) {
+          async function () {
             debug('before(): initialize code callers');
-            codeCallers.init();
-            callback(null);
+            await codeCaller.init();
           },
           async () => {
             debug('before(): start server');
@@ -165,7 +164,7 @@ module.exports = {
       [
         function (callback) {
           debug('after(): finish workers');
-          codeCallers.finish((err) => {
+          codeCaller.finish((err) => {
             if (ERR(err, callback)) return;
             callback(null);
           });
